Handle new Phi-2 tensor names (#104)
Add support for the new set of tensor names used by Phi 2. Unfortunately
does not currently allow original-flavor and new-style to be mixed. I'll
revisit this in the near future when I get to refactoring the
architecture system.

Thanks to whoever at Microsoft changed the `model_type` field - it's
super helpful to have such a clear call-out for the change.