Activation based merging - copied over from wip-zipit branch (#365)
# What is this? 
This PR introduces a way to merge two models via their activations and
hidden states on a tiny sample of data.
This method uses these activations and hidden states to form correlation
matrices to then generate permutation and inverse permutation matrices
for weights in each model and then combines them

This PR consists of three main scripts
1. the first one generates the activation/hidden state for each space
2. a permutation and inverse permutation pair is generated for each
space
3. based on each space and the connected weights, the permutation and/or
inverse permutation is applied to each weight and then the weights are
combined

# Assumptions
The models to be merged are of the same architecture and equal
block/layer count

# Things that couldn't make into the final PR
on-the-fly handling of models with grouped query attention. This hasn't
been tested enough for this release but will be in the near future. For
now, users will have to resort to using this script first:

## Note:
Because this was copied over from another branch (`wip-zipit`) @shamanez
's contributions to the PR is missing, so this is explicit
acknowledgement that @shamanez has worked on this PR alongside other
authors