Add NearSwap (#480)
# NearSwap Algorithm

NearSwap retains most of the weights of the base model, but when a
weight is similar between the two, it is interpolated to the secondary
model value. A parameter t specifies the sameness threshold. When the
distance between two values is below t, the weight from the secondary
model is used.

This PR implements the NearSwap algorithm from
[here](https://huggingface.co/alchemonaut/QuartetAnemoi-70B-t0.0001)

---------

Co-authored-by: Elliot Stein <elliot.stein17@gmail.com>
Co-authored-by: Charles O. Goddard <chargoddard@gmail.com>